chore: move endpoint to POST /functions/deployments/bundle
