+++ packages/cli/lib/zeroYaml/definitions.ts
@@ -33,7 +33,7 @@ import type {
     NangoYamlParsedIntegration,
     ParsedNangoAction,
     ParsedNangoSync,
-    PostFunctionDeploy,
+    PostFunctionDeploymentBundle,
     Result
 } from '@nangohq/types';
 import type * as z from 'zod';
@@ -48,7 +48,7 @@ interface FunctionDefinition {
     data?: { models?: Record<string, ZodModel>; metadata?: ZodMetadata; checkpoint?: ZodCheckpoint } | undefined;
 }
 
-export type FunctionConfig = Omit<PostFunctionDeploy['Body']['functions'][number], 'fileBody'>;
+export type FunctionConfig = Omit<PostFunctionDeploymentBundle['Body']['functions'][number], 'fileBody'>;
 
 export interface ParsedIntegrationDefinitions extends NangoYamlParsed {
     functions: FunctionConfig[];
