+++ packages/types/lib/index.ts
@@ -108,7 +108,6 @@ export type * from './mcp/api.js';
 export type * from './mfa/api.js';
 export type * from './mfa/db.js';
 export type * from './function/config.js';
-export type * from './function/api.js';
 export type * from './function/db.js';
 export type * from './functions/api.js';
 export type * from './functions/domain.js';
